End of Advance Relationship Section

+++ API/Services/CharacterService/CharacterService.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using API.Data;
 using API.Dtos.Character;
 using API.Models;
 using AutoMapper;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace API.Services.CharacterService
@@ -19,22 +21,33 @@ namespace API.Services.CharacterService
         };
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(
+            IMapper mapper,
+            DataContext context,
+            IHttpContextAccessor httpContextAccessor
+        )
         {
+            this._httpContextAccessor = httpContextAccessor;
             this._context = context;
             this._mapper = mapper;
         }
-
+        private int GetUserId()=>int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(
             AddCharacterDto newCharacter
         )
         {
-            Character character = _mapper.Map<Character>(newCharacter);
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
+            Character character = _mapper.Map<Character>(newCharacter);
+            character.User=await _context.Users.FirstOrDefaultAsync(c=>c.Id==GetUserId());
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
-            serviceResponse.Data =await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+            serviceResponse.Data = await _context.Characters
+            .Where(c=>c.User.Id==GetUserId())
+                .Select(c => _mapper.Map<GetCharacterDto>(c))
+                .ToListAsync();
+                
             return serviceResponse;
         }
 
@@ -43,36 +56,53 @@ namespace API.Services.CharacterService
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character =await _context.Characters.FirstAsync(c => c.Id == id);
+                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id&&c.User.Id==GetUserId());
+                if(character !=null)
+                {
                 _context.Characters.Remove(character);
                 await _context.SaveChangesAsync();
-                serviceResponse.Data =await _context.Characters
+                serviceResponse.Data = await _context.Characters
+                .Where(c=>c.User.Id==GetUserId())
                     .Select(c => _mapper.Map<GetCharacterDto>(c))
                     .ToListAsync();
-                ;
+                }
+                else
+                {
+                    serviceResponse.Success=false;
+                    serviceResponse.Message="Character not found";
+                }
             }
             catch (Exception ex)
             {
                 serviceResponse.Success = false;
                 serviceResponse.Message = ex.Message;
             }
+
             return serviceResponse;
         }
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters=await _context.Characters.ToListAsync();
-            serviceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
-            
+            var dbCharacters = await _context.Characters
+            .Include(c=>c.Weapon)
+            .Include(c=>c.Skills)
+            .Where(c=>c.User.Id==GetUserId()).ToListAsync();
+            serviceResponse.Data = dbCharacters
+                .Select(c => _mapper.Map<GetCharacterDto>(c))
+                .ToList();
+
             return serviceResponse;
         }
 
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter=await _context.Characters.FirstOrDefaultAsync(c=>c.Id==id);
-            serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
+            var dbCharacter = await _context.Characters
+            .Include(c=>c.Weapon)
+            .Include(c=>c.Skills)
+            .FirstOrDefaultAsync(c => c.Id == id&&c.User.Id==GetUserId());
+            serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter); 
             return serviceResponse;
         }
 
@@ -83,8 +113,12 @@ namespace API.Services.CharacterService
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
             try
             {
-                Character character =await _context.Characters.FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
-
+                Character character = await _context.Characters
+                .Include(c=>c.User)
+                .FirstOrDefaultAsync(
+                    c => c.Id == updateCharacter.Id
+                );
+                if(character.User.Id==GetUserId()){
                 character.Name = updateCharacter.Name;
                 character.HitPoints = updateCharacter.HitPoints;
                 character.Strength = updateCharacter.Strength;
@@ -94,6 +128,12 @@ namespace API.Services.CharacterService
                 await _context.SaveChangesAsync();
                 serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
             }
+            else
+            {
+                serviceResponse.Success=false;
+                serviceResponse.Message="Character not found";
+            }
+            }
             catch (Exception ex)
             {
                 serviceResponse.Success = false;
@@ -101,5 +141,42 @@ namespace API.Services.CharacterService
             }
             return serviceResponse;
         }
+
+        public async Task<ServiceResponse<GetCharacterDto>> AddCharacterSkill(AddCharacterSkillDto newCharacterSkill)
+        {
+            var response = new ServiceResponse<GetCharacterDto>();
+            try
+            {
+                var character=await _context.Characters
+                .Include(c=>c.Weapon)
+                .Include(c=>c.Skills)
+                .FirstOrDefaultAsync(c=>c.Id==newCharacterSkill.CharacterId&&c.User.Id==GetUserId());
+
+                if(character==null)
+                {
+                    response.Success=false;
+                    response.Message="Character not found";
+                    return response;
+                }
+                var skill=await _context.Skills.FirstOrDefaultAsync(c=>c.Id==newCharacterSkill.SkillId);
+                if (skill==null)
+                {
+                    response.Success=false;
+                    response.Message="Skill not found";
+                    return response;
+                }
+                character.Skills.Add(skill);
+                await _context.SaveChangesAsync();
+                response.Data=_mapper.Map<GetCharacterDto>(character);
+            }
+            catch (Exception ex)
+            {
+                response.Success=false;
+                response.Message=ex.Message;
+            }
+            return response;
+        }
+
+       
     }
 }
